Adicionando funcionalidades aos servicos REST do sistema

+++ src/main/java/com/achieve/carminp/core/rest/re/app/AutorResource.java
@@ -1,12 +1,10 @@
-/**
- * 
- */
 package com.achieve.carminp.core.rest.re.app;
 
 import java.net.URI;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -24,6 +22,9 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.Response.Status;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.achieve.carminp.core.business.in.service.IAutorService;
 import com.achieve.carminp.core.model.im.entidade.AutorEntidade;
 
@@ -35,21 +36,22 @@ import com.achieve.carminp.core.model.im.entidade.AutorEntidade;
  * @since 09/2014
  * @version 1.0
  */
-@RequestScoped
 @Path("/autor")
+@Consumes(MediaType.APPLICATION_JSON)
+@Produces(MediaType.APPLICATION_JSON)
 public class AutorResource {
+	
+	private static final Logger LOGGER = LoggerFactory.getLogger(AutorResource.class.getName());
 	
 	@Inject
 	IAutorService service;
 
 	@POST
-	@Consumes(MediaType.APPLICATION_JSON)
-	@Produces(MediaType.APPLICATION_JSON)
 	public Response criarAutor(@Valid AutorEntidade autor, @Context HttpServletRequest req) {
 		try {
 			service.save(autor);
 		} catch (Exception e) {
-			throw new WebApplicationException(Status.CONFLICT);
+			throw new WebApplicationException(Status.NOT_FOUND);
 		}
 		
 		URI uri = UriBuilder.fromPath("autor/{nome}").build(
@@ -59,18 +61,27 @@ public class AutorResource {
 	}
 	
 	@GET
-	@Path("{id:[0-9][0-9]*}")
-	@Produces(MediaType.APPLICATION_JSON)
-	public AutorEntidade buscarAutorPorId(@PathParam("id")final Long idAutor) {
+	@Path("/{id:[0-9][0-9]*}")
+	public AutorEntidade buscarAutorPorId(@PathParam("id") Long idAutor) {
 		AutorEntidade autorEncontrado = service.getById(idAutor);
-		if (autorEncontrado != null)
+		if (autorEncontrado != null) 
 			return autorEncontrado;
 		
 		throw new WebApplicationException(Status.NOT_FOUND);
 	}
 	
 	@GET
-	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/{nome}")
+	public List<AutorEntidade> buscarAutorPorNome(@PathParam("nome") String nomeAutor) {
+		Map<String, Object> field = new HashMap<String, Object>();
+		field.put(nomeAutor, "autor_nome");
+		
+		List<AutorEntidade> autoresEncontrados = service.findByFields(field, true, 0, null);
+		
+		return autoresEncontrados;
+	}
+	
+	@GET
 	public List<AutorEntidade> buscarTodosAutores() {
 		List<AutorEntidade> autoresEncontrados = service.findAll();
 		
@@ -78,8 +89,13 @@ public class AutorResource {
 	}
 	
 	@DELETE
-	@Path("{id:[0-9][0-9]*}")
-	public void removerAutor(@PathParam("id")final Long idAutor) {
-		service.delete(idAutor);
+	@Path("/{id:[0-9][0-9]*}")
+	public void removerAutor(@PathParam("id") Long idAutor) {
+		if (idAutor != null) {
+			LOGGER.info("Removendo autor com id {}", idAutor);
+			service.delete(idAutor);
+		} else {
+			LOGGER.info("Autor com id {} não existe e, portanto, nada foi excluído", idAutor);
+		}
 	}
 }
